Add structured logger middleware for Gin with request ID and sensitive data masking
- Implemented Logger middleware to log request details including request ID, IP, method, path, and user agent.
- Added functionality to sanitize URLs by masking sensitive query parameters.
- Created utility functions to retrieve logger and request ID from Gin context.
- Included context management for logger to ensure consistent logging across requests.
